+++ Services/ExpenseService.cs
@@ -2,7 +2,7 @@
 {
     internal class ExpenseService : IExpenseService
     {
-        public string Add(string description, int amount)
+        public string Add(string description, double amount)
         {
             throw new NotImplementedException();
         }
